Q4
Write down a C++ class statement to represent a person, containing fields to represent
the person's first and last names. Demonstrate how to declare a constructor and a
destructor. Declare an accessor method called get name to return the person's name.
Declare a Student as a subclass of Person. What access level is required on the fields
of Person so that they are accessible to Student but not to external code?